+++ clients/admin-ui/src/features/messaging-templates/messaging-templates.slice.ts
@@ -1,8 +1,5 @@
 import { baseApi } from "~/features/common/api.slice";
-import {
-  MinimalProperty,
-  Page_MessagingTemplateWithPropertiesSummary_,
-} from "~/types/api";
+import { MinimalProperty } from "~/types/api";
 import { BulkUpdateFailed } from "~/types/api/models/BulkUpdateFailed";
 
 export type MessagingTemplate = {
@@ -70,17 +67,6 @@ const messagingTemplatesApi = baseApi.injectEndpoints({
       query: () => ({ url: `messaging/templates` }),
       providesTags: () => ["Messaging Templates"],
     }),
-    getSummaryMessagingTemplates: build.query<
-      Page_MessagingTemplateWithPropertiesSummary_,
-      any
-    >({
-      query: (params) => ({
-        method: "GET",
-        url: `messaging/templates/summary`,
-        params,
-      }),
-      providesTags: () => ["Property-Specific Messaging Templates"],
-    }),
     updateMessagingTemplates: build.mutation<
       BulkPutMessagingTemplateResponse,
       MessagingTemplate[]
@@ -90,7 +76,7 @@ const messagingTemplatesApi = baseApi.injectEndpoints({
         method: "PUT",
         body: templates,
       }),
-      invalidatesTags: () => ["Property-Specific Messaging Templates"],
+      invalidatesTags: () => ["Messaging Templates"],
     }),
     // Render data from existing template- GET by id
     getMessagingTemplateById: build.query<MessagingTemplateResponse, string>({
@@ -98,30 +84,6 @@ const messagingTemplatesApi = baseApi.injectEndpoints({
         url: `/messaging/templates/${templateId}`,
       }),
       providesTags: () => ["Property-Specific Messaging Templates"],
-    }),
-    // Update existing template
-    putMessagingTemplateById: build.mutation<
-      MessagingTemplateResponse,
-      MessagingTemplateUpdate
-    >({
-      query: ({ templateId, template }) => ({
-        url: `/messaging/templates/${templateId}`,
-        method: "PUT",
-        body: template,
-      }),
-      invalidatesTags: () => ["Property-Specific Messaging Templates"],
-    }),
-    // Update existing template
-    patchMessagingTemplateById: build.mutation<
-      MessagingTemplateResponse,
-      MessagingTemplatePatch
-    >({
-      query: ({ templateId, template }) => ({
-        url: `/messaging/templates/${templateId}`,
-        method: "PATCH",
-        body: template,
-      }),
-      invalidatesTags: () => ["Property-Specific Messaging Templates"],
     }),
     // endpoint for rendering data for default template- GET by type
     getMessagingTemplateDefault: build.query<
@@ -131,18 +93,6 @@ const messagingTemplatesApi = baseApi.injectEndpoints({
       query: (templateType: string) => ({
         url: `/messaging/templates/default/${templateType}`,
       }),
-    }),
-    // endpoint for creating new messaging template- POST by type
-    createMessagingTemplateByType: build.mutation<
-      MessagingTemplateResponse,
-      MessagingTemplateCreate
-    >({
-      query: ({ templateType, template }) => ({
-        url: `/messaging/templates/${templateType}`,
-        method: "POST",
-        body: template,
-      }),
-      invalidatesTags: () => ["Property-Specific Messaging Templates"],
     }),
     // delete template by id
     deleteMessagingTemplateById: build.mutation<void, string>({
@@ -158,11 +108,7 @@ const messagingTemplatesApi = baseApi.injectEndpoints({
 export const {
   useGetMessagingTemplatesQuery,
   useUpdateMessagingTemplatesMutation,
-  useGetSummaryMessagingTemplatesQuery,
   useGetMessagingTemplateByIdQuery,
-  usePutMessagingTemplateByIdMutation,
   useGetMessagingTemplateDefaultQuery,
-  useCreateMessagingTemplateByTypeMutation,
   useDeleteMessagingTemplateByIdMutation,
-  usePatchMessagingTemplateByIdMutation,
 } = messagingTemplatesApi;